+++ index.js
@@ -50,6 +50,8 @@ import "./features/chat/messageOnRevenant";
 import "./features/chat/messageOnHotspotFound";
 import "./features/chat/compactCatchMessages";
 
+import "./features/world/boxEntities";
+
 register('worldLoad', () => {
     Client.showTitle('', '', 1, 1, 1); // Shitty fix for a title not showing for the 1st time
 });